Functionality for posterior correlation matrix
