+++ eeyore/stats/__init__.py
@@ -1,10 +1,14 @@
-from .cov import cov, recursive_cov
+from .cor import cor
+from .cor_from_cov import cor_from_cov
+from .cov import cov
 from .discrepancy import mmd
 from .inse_mc_cov import inse_mc_cov
 from .loss import binary_cross_entropy
+from .mc_cor import mc_cor
 from .mc_cov import mc_cov
-from .mean import recursive_mean
 from .metrics import softabs
 from .multi_ess import multi_ess
 from .multi_rhat import multi_rhat
 from .random import choose, choose_from_subset
+from .recursive_cov import recursive_cov
+from .recursive_mean import recursive_mean
