+++ eeyore/stats/recursive_mean.py
